+++ metrics.go
@@ -5,30 +5,39 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-func metrics(namespace string, subsystem string, result runResult) *prometheus.Registry {
+func metrics(namespace string, subsystem string, probe probeType, result runResult) *prometheus.Registry {
 	// Init metrics
 	log.Debug("[metrics] Initialize metrics: ", subsystem)
 
-	gaugeUp := prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: subsystem,
-		Name:      "up",
-		Help:      "General availability of this probe",
-	})
+	gaugeUp := prometheus.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "up",
+			Help:      "General availability of this probe",
+		},
+		probe.labelNames,
+	)
 
-	gaugeSuccess := prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: subsystem,
-		Name:      "success",
-		Help:      "Show if the script was executed successfully",
-	})
+	gaugeSuccess := prometheus.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "success",
+			Help:      "Show if the script was executed successfully",
+		},
+		probe.labelNames,
+	)
 
-	gaugeDuration := prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: subsystem,
-		Name:      "duration_seconds",
-		Help:      "Shows the execution time of the script",
-	})
+	gaugeDuration := prometheus.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "duration_seconds",
+			Help:      "Shows the execution time of the script",
+		},
+		probe.labelNames,
+	)
 
 	// Register metrics
 	log.Debug("[metrics] Register metrics: ", subsystem)
@@ -41,15 +50,15 @@ func metrics(namespace string, subsystem string, result runResult) *prometheus.R
 	// Set metrics
 	log.Debug("[metrics] Set metrics: ", subsystem)
 
-	gaugeUp.Set(1)
+	gaugeUp.WithLabelValues(probe.labelValues...).Set(1)
 
 	if result.exitCode == 0 {
-		gaugeSuccess.Set(1)
+		gaugeSuccess.WithLabelValues(probe.labelValues...).Set(1)
 	} else {
-		gaugeSuccess.Set(0)
+		gaugeSuccess.WithLabelValues(probe.labelValues...).Set(0)
 	}
 
-	gaugeDuration.Set(result.duration.Seconds())
+	gaugeDuration.WithLabelValues(probe.labelValues...).Set(result.duration.Seconds())
 
 	return registry
 }